requirements.txt
Baker==1.3
Jinja2==2.7.1
MarkupSafe==0.18
coverage==3.7.1
nose==1.3.0
python-dateutil==2.2
six==1.4.1
unicodecsv==0.9.4
wsgiref==0.1.2
